requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
pyotp==2.9.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.5
python-dotenv==1.0.1
blinker==1.7.0
colorama==0.4.6
gunicorn==21.2.0
packaging==23.2
pytz==2024.1
PyGithub==1.59.1
